+++ lib/maru/types/charlist.ex
@@ -1,6 +1,6 @@
-defmodule Maru.Types.CharList do
+defmodule Maru.Types.Charlist do
   @moduledoc """
-  Buildin Type: CharList
+  Buildin Type: Charlist
   """
 
   use Maru.Type
@@ -8,7 +8,7 @@ defmodule Maru.Types.CharList do
   @doc false
   def parse(input, _) when is_list(input), do: input
   def parse(input, _) do
-    input |> to_char_list
+    input |> to_charlist
   end
 
 end