+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Enhancements
   * support set version by `accept` header
   * refactor unittest
+  * return original value for `map` and `list` type without `do block`
 * Deprecations
   * do NOT support elixir 1.2.x any more